requirements.txt
e2b-code-interpreter==1.0.4
anthropic==0.39.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3